+++ src/domain/community/organization-role/organization.role.module.ts
@@ -1,0 +1,26 @@
+import { Module } from '@nestjs/common';
+import { OrganizationRoleService } from './organization.role.service';
+import { OrganizationRoleResolverMutations } from './organization.role.resolver.mutations';
+import { OrganizationRoleResolverFields } from './organization.role.resolver.fields';
+import { UserModule } from '@domain/community/user/user.module';
+import { AuthorizationModule } from '@core/authorization/authorization.module';
+import { AgentModule } from '@domain/agent/agent/agent.module';
+import { AuthorizationPolicyModule } from '@domain/common/authorization-policy/authorization.policy.module';
+import { ContributorLookupModule } from '@services/infrastructure/contributor-lookup/contributor.lookup.module';
+
+@Module({
+  imports: [
+    AgentModule,
+    AuthorizationPolicyModule,
+    AuthorizationModule,
+    ContributorLookupModule,
+    UserModule,
+  ],
+  providers: [
+    OrganizationRoleService,
+    OrganizationRoleResolverMutations,
+    OrganizationRoleResolverFields,
+  ],
+  exports: [OrganizationRoleService],
+})
+export class OrganizationRoleModule {}
