+++ src/domain/community/organization-role/organization.role.resolver.fields.ts
@@ -1,0 +1,108 @@
+import { UseGuards } from '@nestjs/common';
+import { Resolver } from '@nestjs/graphql';
+import { Parent, ResolveField } from '@nestjs/graphql';
+import { OrganizationRoleService } from './organization.role.service';
+import { AuthorizationPrivilege } from '@common/enums';
+import { GraphqlGuard } from '@core/authorization';
+import { IOrganization, Organization } from '@domain/community/organization';
+import { IUser } from '@domain/community/user';
+import { CurrentUser, Profiling } from '@common/decorators';
+import { AgentInfo } from '@core/authentication.agent.info/agent.info';
+import { AuthorizationService } from '@core/authorization/authorization.service';
+import { OrganizationRole } from '@common/enums/organization.role';
+import { OrganizationService } from '../organization/organization.service';
+
+@Resolver(() => IOrganization)
+export class OrganizationRoleResolverFields {
+  constructor(
+    private authorizationService: AuthorizationService,
+    private organizationService: OrganizationService,
+    private organizationRoleService: OrganizationRoleService
+  ) {}
+
+  @UseGuards(GraphqlGuard)
+  @ResolveField('associates', () => [IUser], {
+    nullable: true,
+    description: 'All Users that are associated with this Organization.',
+  })
+  @Profiling.api
+  async associates(
+    @Parent() parent: Organization,
+    @CurrentUser() agentInfo: AgentInfo
+  ): Promise<IUser[]> {
+    // Reload to ensure the authorization is loaded
+    const organization = await this.organizationService.getOrganizationOrFail(
+      parent.id
+    );
+    await this.authorizationService.grantAccessOrFail(
+      agentInfo,
+      organization.authorization,
+      AuthorizationPrivilege.READ,
+      `read associates on org: ${organization.nameID}`
+    );
+
+    return await this.organizationRoleService.getAssociates(organization);
+  }
+
+  @UseGuards(GraphqlGuard)
+  @ResolveField('admins', () => [IUser], {
+    nullable: true,
+    description: 'All Users that are admins of this Organization.',
+  })
+  @Profiling.api
+  async admins(
+    @Parent() parent: Organization,
+    @CurrentUser() agentInfo: AgentInfo
+  ): Promise<IUser[]> {
+    // Reload to ensure the authorization is loaded
+    const organization = await this.organizationService.getOrganizationOrFail(
+      parent.id
+    );
+    await this.authorizationService.grantAccessOrFail(
+      agentInfo,
+      organization.authorization,
+      AuthorizationPrivilege.READ,
+      `read admins on org: ${organization.nameID}`
+    );
+
+    return await this.organizationRoleService.getAdmins(organization);
+  }
+
+  @UseGuards(GraphqlGuard)
+  @ResolveField('owners', () => [IUser], {
+    nullable: true,
+    description: 'All Users that are owners of this Organization.',
+  })
+  @Profiling.api
+  async owners(
+    @Parent() parent: Organization,
+    @CurrentUser() agentInfo: AgentInfo
+  ): Promise<IUser[]> {
+    // Reload to ensure the authorization is loaded
+    const organization = await this.organizationService.getOrganizationOrFail(
+      parent.id
+    );
+    await this.authorizationService.grantAccessOrFail(
+      agentInfo,
+      organization.authorization,
+      AuthorizationPrivilege.READ,
+      `read owners on org: ${organization.nameID}`
+    );
+
+    return await this.organizationRoleService.getOwners(organization);
+  }
+
+  @UseGuards(GraphqlGuard)
+  @ResolveField('myRoles', () => [OrganizationRole], {
+    nullable: true,
+    description:
+      'The roles on this Organization for the currently logged in user.',
+  })
+  @Profiling.api
+  async myRoles(
+    @CurrentUser() agentInfo: AgentInfo,
+    @Parent() organization: IOrganization
+  ): Promise<OrganizationRole[]> {
+    return this.organizationRoleService.getMyRoles(agentInfo, organization);
+  }
+}
