+++ src/domain/community/organization/organization.resolver.mutations.ts
@@ -12,7 +12,6 @@ import { GraphqlGuard } from '@core/authorization';
 import { AuthorizationPrivilege } from '@common/enums';
 import { OrganizationAuthorizationService } from './organization.service.authorization';
 import { AgentInfo } from '@core/authentication.agent.info/agent.info';
-import { IUser } from '@domain/community/user/user.interface';
 import { OrganizationAuthorizationResetInput } from './dto/organization.dto.reset.authorization';
 import { UserGroupAuthorizationService } from '../user-group/user-group.service.authorization';
 import { AuthorizationService } from '@core/authorization/authorization.service';
@@ -23,9 +22,6 @@ import { PreferenceSetService } from '@domain/common/preference-set/preference.s
 import { CreateUserGroupInput } from '../user-group/dto';
 import { PlatformAuthorizationPolicyService } from '@src/platform/authorization/platform.authorization.policy.service';
 import { IOrganization } from './organization.interface';
-import { RemoveOrganizationRoleFromUserInput } from './dto/organization.dto.remove.role.from.user';
-import { AssignOrganizationRoleToUserInput } from './dto/organization.dto.assign.role.to.user';
-import { OrganizationRole } from '@common/enums/organization.role';
 
 @Resolver(() => IOrganization)
 export class OrganizationResolverMutations {
@@ -164,64 +160,6 @@ export class OrganizationResolverMutations {
     );
     return await this.organizationAuthorizationService.applyAuthorizationPolicy(
       organization
-    );
-  }
-
-  @UseGuards(GraphqlGuard)
-  @Mutation(() => IUser, {
-    description: 'Assigns an Organization Role to user.',
-  })
-  @Profiling.api
-  async assignOrganizationRoleToUser(
-    @CurrentUser() agentInfo: AgentInfo,
-    @Args('membershipData') membershipData: AssignOrganizationRoleToUserInput
-  ): Promise<IUser> {
-    const organization = await this.organizationService.getOrganizationOrFail(
-      membershipData.organizationID
-    );
-
-    await this.authorizationService.grantAccessOrFail(
-      agentInfo,
-      organization.authorization,
-      AuthorizationPrivilege.GRANT,
-      `assign organization role to user: ${organization.nameID} - ${membershipData.role}`
-    );
-    return await this.organizationService.assignOrganizationRoleToUser(
-      membershipData
-    );
-  }
-
-  @UseGuards(GraphqlGuard)
-  @Mutation(() => IUser, {
-    description: 'Removes Organization Role from user.',
-  })
-  @Profiling.api
-  async removeOrganizationRoleFromUser(
-    @CurrentUser() agentInfo: AgentInfo,
-    @Args('membershipData') membershipData: RemoveOrganizationRoleFromUserInput
-  ): Promise<IUser> {
-    const organization = await this.organizationService.getOrganizationOrFail(
-      membershipData.organizationID
-    );
-    let authorization = organization.authorization;
-    if (membershipData.role === OrganizationRole.ASSOCIATE) {
-      // Extend the authorization policy with a credential rule to assign the GRANT privilege
-      // to the user specified in the incoming mutation. Then if it is the same user as is logged
-      // in then the user will have the GRANT privilege + so can carry out the mutation
-      authorization =
-        this.organizationAuthorizationService.extendAuthorizationPolicyForSelfRemoval(
-          organization,
-          membershipData.userID
-        );
-    }
-    await this.authorizationService.grantAccessOrFail(
-      agentInfo,
-      authorization,
-      AuthorizationPrivilege.GRANT,
-      `remove user organization role from user: ${organization.nameID} - ${membershipData.role}`
-    );
-    return await this.organizationService.removeOrganizationRoleFromUser(
-      membershipData
     );
   }
